Drop go-1.6 code
