+++ lxd/db/query/slices_test.go
@@ -23,6 +23,15 @@ func TestStrings_Error(t *testing.T) {
 	}
 }
 
+var testStringsErrorCases = []struct {
+	query string
+	error string
+}{
+	{"garbage", "near \"garbage\": syntax error"},
+	{"SELECT id, name FROM test", "query yields 2 columns, not 1"},
+	{"SELECT id FROM test", "query yields INTEGER column, not TEXT"},
+}
+
 // All values yield by the query are returned.
 func TestStrings(t *testing.T) {
 	tx := newTxForSlices(t)
@@ -41,6 +50,15 @@ func TestIntegers_Error(t *testing.T) {
 			assert.Nil(t, values)
 		})
 	}
+}
+
+var testIntegersErrorCases = []struct {
+	query string
+	error string
+}{
+	{"garbage", "near \"garbage\": syntax error"},
+	{"SELECT id, name FROM test", "query yields 2 columns, not 1"},
+	{"SELECT name FROM test", "query yields TEXT column, not INTEGER"},
 }
 
 // All values yield by the query are returned.
